+++ httputil/middleware_example_test.go
@@ -1,4 +1,4 @@
-package http
+package httputil
 
 import (
 	"fmt"
